+++ lib/src/discover/discover.dart
@@ -1,0 +1,3 @@
+export 'codegen.dart';
+export 'models.dart';
+export 'scanner.dart';